WeChat：myBatisPlus 替代 myBatis ，没写完没测试，主要更改在数据层，更改了包名接口名

+++ src/main/java/com/jeffrey/wechat/service/UserQuestionService.java
@@ -24,6 +24,6 @@ public interface UserQuestionService {
      * @param openid 用户唯一 id
      * @return
      */
-    int feedBackIsExists(String openid);
+    Long feedBackIsExists(String openid);
 
 }
